% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytetype.R
\name{CleanUpArtifacts}
\alias{CleanUpArtifacts}
\title{Clean Up Artifacts}
\usage{
CleanUpArtifacts(prepped_data)
}
\arguments{
\item{prepped_data}{Named list containing prepared data from \code{PrepareCyteTypeR()}.}
}
\description{
Cleans up the artifact files after the run.
}
