% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{cytetype_api_error}
\alias{cytetype_api_error}
\title{Custom API Error}
\usage{
cytetype_api_error(message, call = NULL)
}
\arguments{
\item{message}{Error message}

\item{call}{Call that generated the error}
}
\value{
An error object of class cytetype_api_error
}
\description{
Custom API Error
}
